requirements.txt
azure-identity==1.15
azure-keyvault-keys==4.3
azure-keyvault-secrets==4.8